pyproject.toml
[project]
name = "django-shop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "coverage>=7.12.0",
    "django>=5.2.7",
    "django-debug-toolbar>=6.1.0",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.1",
    "faker>=37.11.0",
    "polars>=1.34.0",
    "psycopg[binary]>=3.2.10",
    "pytest>=9.0.1",
    "pytest-django>=4.11.1",
    "pytest-html>=4.1.1",
    "tqdm>=4.67.1",
]
